Added TODO comment
From what I understand, the isCargoAtSensor method returns the value of the beam break sensor that is at the top of the shooting tower. For the receiving method, we need the beam break at the bottom of the tower, which I'm not yet sure if it is going to be initialized in the indexer or intake subsystem.