Change to shoot when in ready to shoot state
Was just checking cargo at sensor which caused us to try to
shoot too soon. The state machine would force the proper
state sequence thus causing the shot to not happen. The
fix is to wait until the proper state is reached. BTW, the
shoot trigger used in autonomous was already doin this
and was working fine.