Be in manual mode before initializing manual cmds
The initialization of the commands was causing end to be called on the current
state command. This is okay, but because the manual mode flag was not
set yet, the state command's end schedules the next state via
PostLoopCommandScheduler. When post loop trys to really schedule that
next state, it causes the manual command that just started this whole
thing off to be cancelled. Therefore, its end is called, which tries to
schedule a state command via PostLoopCommandScheduler leading
to the concurrent modification exception that we see. We were still
iterating the list of commands to start, when this last schedule was
attempted.